HRCPP-494 Removed TcpTransportFactory hierarchy
