+++ src/hotrod/impl/transport/tcp/TcpTransport.cpp
@@ -1,6 +1,6 @@
 #include "infinispan/hotrod/exceptions.h"
 #include "hotrod/impl/transport/tcp/TcpTransport.h"
-#include "hotrod/impl/transport/tcp/TcpTransportFactory.h"
+#include "hotrod/impl/transport/TransportFactory.h"
 #include "hotrod/sys/Socket.h"
 
 #include <iostream>
