--- src/hotrod/impl/transport/tcp/TcpTransportFactory.h
@@ -1,79 +1,0 @@
-#ifndef ISPN_HOTROD_TRANSPORT_TCPTRANSPORTFACTORY_H
-#define ISPN_HOTROD_TRANSPORT_TCPTRANSPORTFACTORY_H
-
-
-#include "infinispan/hotrod/Configuration.h"
-#include "hotrod/impl/transport/Transport.h"
-#include "hotrod/impl/transport/TransportFactory.h"
-#include "hotrod/impl/transport/tcp/ConnectionPool.h"
-#include "hotrod/impl/transport/tcp/TcpTransport.h"
-#include "hotrod/impl/transport/tcp/TransportObjectFactory.h"
-#include "hotrod/impl/consistenthash/ConsistentHashFactory.h"
-#include "hotrod/sys/Mutex.h"
-
-#include <vector>
-
-namespace infinispan {
-namespace hotrod {
-namespace transport {
-
-class RequestBalancingStrategy;
-class InetSocketAddress;
-
-class TcpTransportFactory : public TransportFactory
-{
-  public:
-    TcpTransportFactory(const Configuration& config) : TransportFactory(TopologyInfo(initialServers, config)), configuration(config), maxRetries(config.getMaxRetries()) {}
-    void start(protocol::Codec& codec, ClientListenerNotifier* );
-    void destroy();
-
-    transport::Transport& getTransport(const std::vector<char>& cacheName, const std::set<transport::InetSocketAddress>& failedServers);
-    transport::Transport& getTransport(const std::vector<char>& key, const std::vector<char>& cacheName, const std::set<transport::InetSocketAddress>& failedServers);
-
-    void releaseTransport(Transport& transport);
-    void invalidateTransport(
-        const InetSocketAddress& address, Transport* transport);
-    bool clusterSwitch();
-    bool clusterSwitch(std::string clusterName);
-    bool isTcpNoDelay();
-    int getMaxRetries();
-    int getSoTimeout();
-    int getConnectTimeout();
-    const std::string& getSniHostName();
-    bool isSslEnabled();
-    const std::string& getSslServerCAPath();
-    const std::string& getSslServerCAFile();
-    const std::string& getSslClientCertificateFile();
-
-    void updateServers(std::vector<InetSocketAddress>& );
-    void updateHashFunction(
-            std::vector<std::vector<InetSocketAddress>>& segmentOwners,
-            uint32_t &numSegment, uint8_t &hashFunctionVersion,
-            const std::vector<char>& cacheName, int topologyId);
-    Transport& borrowTransportFromPool(const InetSocketAddress& server);
-    virtual ~TcpTransportFactory() { }
-    const Configuration& getConfiguration() { return configuration; }
-
-  private:
-    sys::Mutex lock;
-    std::vector<InetSocketAddress> initialServers;
-    const Configuration& configuration;
-    int maxRetries;
-    std::shared_ptr<TransportObjectFactory> transportFactory;
-    std::shared_ptr<ConnectionPool> connectionPool;
-    std::shared_ptr<FailOverRequestBalancingStrategy> balancer;
-    std::string currCluster;
-    void createAndPreparePool();
-    void updateTransportCount();
-    void pingServers();
-    ConnectionPool* getConnectionPool();
-    std::vector<ServerConfiguration> getNextWorkingServersConfiguration();
-    void pingExternalServer(InetSocketAddress s);
-    std::string sniHostName;
-    ClientListenerNotifier* listenerNotifier;
-};
-
-}}} // namespace infinispan::hotrod::transport
-
-#endif  /* ISPN_HOTROD_TRANSPORT_TCPTRANSPORTFACTORY_H */
-
